Fitting the cumulative incidence function for time-to event data under ICH E9 (R1)

This function estimates the potential cumulative incidence function for time-to event data under ICH E9 (R1) to address intercurrent events. Multiple strategies are allowed. The input data should be of semicompeting risks structure.

scr.ICH(
  A,
  Time,
  status,
  Time_int,
  status_int,
  strategy = "composite",
  cov1 = NULL,
  method = "np",
  weights = NULL,
  subset = NULL
)

Arguments

A

Treatment indicator, 1 for treatment and 0 for control.

Time

Time to the primary (terminal) event.

status

Indicator of the primary (terminal) event, 1 for event and 0 for censoring.

Time_int

Time to the intercurrent event.

status_int

Indicator of the intercurrent event, 1 for event and 0 for censoring.

strategy

Stragety to address intercurrent events, "treatment" indicating treatment policy strategy, "composite" indicating composite variable strategy, "natural" indicating hypothetical strategy (Scenario I, controlling the hazard of intercurrent events), "removed" indicating hypothetical strategy (Scenario II, removing intercurrent events), "whileon" indicating while on treatment strategy, and "principal" indicating principal stratum strategy.

cov1

Baseline covariates.

method

Estimation method, "np" indicating nonparametric estimation, otherwise indicating semiparametrically efficient estimation based on efficient influence functions.

weights

Weight for each subject.

subset

Subset, either numerical or logical.

Value

A list including the fitted object and input variables.

Details

Intercurrent events refer to the events occurring after treatment initiation of clinical trials that affect either the interpretation of or the existence of the measurements associated with the clinical question of interest. The International Conference on Harmonization (ICH) E9 (R1) addendum proposed five strategies have been proposed in to address intercurrent events, namely, treatment policy strategy, composite variable strategy, while on treatment strategy, hypothetical strategy, and principal stratum strategy. To answer a specific scientific question, a strategy with a particular estimand is chosen before the study design. \cr We adopt the potential outcomes framework that defines a causal estimand as the contrast between functionals of potential outcomes. Consider a randomized controlled trial with n individuals randomly assigned to one of two treatment conditions, denoted by w, where w = 1 represents the active treatment (a test drug) and w = 0 represents the control (placebo). Assume that all patients adhere to their treatment assignments and do not discontinue treatment. Associated with individual i = 1, ..., n are two potential time-to-event primary outcomes T_i(1) and T_i(0), if any, which represent the time durations from treatment initiation to the primary outcome event under two treatment assignments respectively. Let R_i(1) and R_i(0) denote the occurrence time of potential intercurrent events, if any, under the two treatment assignments, respectively. Intercurrent events are considered as absent if no post-treatment intercurrent events occur until the end of study. \cr We adopt the potential cumulative incidences under both treatment assignments as the target estimands. Potential cumulative incidences describe the probability of time-to-event outcomes occurring at each time point. We define the treatment effect as the contrast of two potential cumulative incidences. Cumulative incidences are model-free and collapsible, enjoying causal interpretations.

See also

Examples

## Generate simulated data
## dat = .generatedata(500)

## composite variable strategy, 
## nonparametric estimation without covariates
fit1 = surv.ICH(dat$Z, dat$Time, dat$cstatus,"composite")

## hypothetical strategy (natural effects), 
## nonparametric estimation with inverse probability weighting
ps = predict(glm(dat$Z ~ dat$X, family='binomial'), type='response')
w = dat$Z/ps + (1-dat$Z)/(1-ps)
fit1 = surv.ICH(dat$Z, dat$Time, dat$cstatus, "natural", dat$X, weights=w)

## composite variable strategy, 
## semiparametrically efficient estimation with covariates
fit2 = scr.ICH(dat$Z, dat$T, dat$dT, dat$R, dat$dR, "composite", dat$X, method='eff')