Fitting the cumulative incidence function using principal stratum strategy

This function estimates the potential cumulative incidence function based on efficient influence functions using principal stratum strategy (competing risks data structure). Cox models are employed for survival models. The estimand is defined in a subpopulation where intercurrent events would never occur regardless of treatment conditions.

surv.principal.eff(A, Time, cstatus, X = NULL, subset = NULL)

Arguments

A

Treatment indicator, 1 for treatment and 0 for control.

Time

Time to event.

cstatus

Indicator of event, 1 for the primary event, 2 for the intercurrent event, 0 for censoring.

X

Baseline covariates.

subset

Subset, either numerical or logical.

Value

A list including

time1

Time points in the treated group.

time0

Time points in the control group.

cif1

Estimated cumulative incidence function in the treated group.

cif0

Estimated cumulative incidence function in the control group.

se1

Standard error of the estimated cumulative incidence function in the treated group.

se0

Standard error of the estimated cumulative incidence function in the control group.

tt

Time points in both groups.

ate

Estimated treatment effect (difference in cumulative incidence functions).

se

Standard error of the estimated treatment effect.

p.val

P value of testing the treatment effect based on the efficient influence function of the restricted mean survival time lost by the end of study.

Details

The principal stratum strategy aims to stratify the population into subpopulations based on the joint potential occurrences of intercurrent events under the two treatment assignments (R(1), R(0)). Suppose we are interested in a principal stratum comprised of individuals who would never experience intercurrent events, regardless of which treatment they receive. This principal stratum can be indicated by \{R(1)=R(0)=\infty\}. The treatment effect is now defined within this subpopulation, \tau(t) = P(T(1) < t \mid R(1)=R(0)=\infty) - P(T(0) < t \mid R(1)=R(0)=\infty), representing the difference in probabilities of experiencing primary outcome events during (0,t) under active treatment and placebo in the subpopulation that will not experience intercurrent events regardless of treatment during (0,t).

See also

Examples

1
#> [1] 1